Move dummy DataGrid content to helper folder
