+++ Source/DesignSystem/molecules/DataGrid/helpers.tsx
@@ -1,12 +1,6 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import React from 'react';
-
-import { GridColDef } from '@mui/x-data-grid-pro';
-
-import { EditCell, EditTextFieldCell, IconButton, StatusIndicator } from '@dolittle/design-system';
-
 export const dataTableDescription = `Data tables are an excellent way to organize vast amounts of data. Data tables should be easy to scan, 
             allowing the user to look for patterns and develop insights.
 
@@ -29,165 +23,3 @@ Only allow for one row to expand at a time so as not to overwhelm the user.
 
 * Row: Rows can be expandable and collapsable offering the opportunity to include more subsequent data beneath. Content in rows should be spaced evenly apart. Text should left align vertically within its column
 and numbers, including dates, should right align for better scanning.`;
-
-type DummyRowsProps = {
-    id: string;
-    col1?: string;
-    col2?: string;
-    col3?: string;
-    col4?: string;
-};
-
-export const dummyRows: DummyRowsProps[] = [
-    { id: '1', col1: 'Row 1', col2: 'Row 1', col3: 'Row 1', col4: 'Row 1' },
-    { id: '2', col1: 'Row 2', col2: 'Row 2', col3: 'Row 2', col4: 'Row 2' },
-    { id: '3', col1: 'Row 3', col2: 'Row 3', col3: 'Row 3', col4: 'Row 3' },
-    { id: '4', col1: 'Row 4', col2: 'Row 4', col3: 'Row 4', col4: 'Row 4' },
-];
-export const dummyColumns: GridColDef[] = [
-    {
-        field: 'col1',
-        headerName: 'Column 1',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-    },
-    {
-        field: 'col2',
-        headerName: 'Column 2',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-    },
-    {
-        field: 'col3',
-        headerName: 'Column 3',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-    },
-    {
-        field: 'col4',
-        headerName: 'Column 4',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-    },
-];
-
-export const dummyIconRows: DummyRowsProps[] = [{ id: '1' }];
-export const dummyIconColumns: GridColDef[] = [
-    {
-        field: 'success',
-        headerName: 'Status success',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        headerAlign: 'center',
-        align: 'center',
-        renderCell: () => <StatusIndicator status='table-success' />,
-    },
-    {
-        field: 'warning',
-        headerName: 'Status warning',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        headerAlign: 'center',
-        align: 'center',
-        renderCell: () => <StatusIndicator status='warning' />,
-    },
-    {
-        field: 'error',
-        headerName: 'Status error',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        headerAlign: 'center',
-        align: 'center',
-        renderCell: () => <StatusIndicator status='error' />,
-    },
-    {
-        field: 'unknown',
-        headerName: 'Status unknown',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        headerAlign: 'center',
-        align: 'center',
-        renderCell: () => <StatusIndicator status='unknown' />,
-    },
-    {
-        field: 'download',
-        headerName: 'Download file',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        headerAlign: 'center',
-        align: 'center',
-        renderCell: () => (
-            <IconButton
-                tooltipText='Download'
-                icon='DownloadRounded'
-                href={URL.createObjectURL(new Blob([], { type: 'text/plain' }))}
-                download='sample.log'
-            />
-        ),
-    },
-];
-
-export const dummyEditCellsColumns: GridColDef[] = [
-    {
-        field: 'col1',
-        headerName: 'Column 1',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        editable: true,
-        renderCell: EditCell,
-        renderEditCell: EditTextFieldCell,
-    },
-    {
-        field: 'col2',
-        headerName: 'Column 2',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        editable: true,
-        renderCell: EditCell,
-        renderEditCell: EditTextFieldCell,
-    },
-    {
-        field: 'col3',
-        headerName: 'Column 3',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        editable: true,
-        renderCell: EditCell,
-        renderEditCell: EditTextFieldCell,
-    },
-    {
-        field: 'col4',
-        headerName: 'Column 4',
-        sortable: false,
-        minWidth: 150,
-        flex: 1,
-        editable: true,
-        renderCell: EditCell,
-        renderEditCell: EditTextFieldCell,
-    },
-];
-
-// TODO: Don't know yet if these are correct or necessary guidelines:
-
-// * Header: Header columns should be adjustable in width size.If a column width must be reduced to a width smaller than the column name, use the appropriate abbreviation for the column label
-// and a tooltip upon hover to indicate the actual name.If abbreviation is not known, truncate the label.
-
-// **Size**: Use medium data table rows by default. When space is at a premium and the data sets are not overly complex, it's okay to use a dense table instead;
-
-// * Dividers: All rows, including the header row, should be separated by a divider.Expanded rows with additional content does not need to include a divider but can
-// if it offers better structure and separation between data sets.
-
-// * Pagination: Pagination can be added when data tables contain enough information that would cause the user to scroll vertically.
-// Do not build a data table that is longer than the height of a standard viewport.Instead, use pagination to break up the table.
